player wasnt dealing well with collisions

+++ include/cub3d.h
@@ -139,7 +139,7 @@ typedef struct s_ray
     t_vector *hit;        // Exact hit position on the wall
     int side;             // Which side of the wall was hit (0 = vertical, 1 = horizontal)
     int tex_x;            // X-coordinate on the texture
-    double lineHeight;    // Height of the line to draw (based on distance to the wall)
+    int lineHeight;    // Height of the line to draw (based on distance to the wall)
     int drawStart;        // Y-coordinate to start drawing (top of the wall)
     int drawEnd;          // Y-coordinate to end drawing (bottom of the wall)
     double dirX;          // Direction vector X component
